Added functionality to the Receiver and the Invoker classes.

+++ CommandTroops/Commands/TurnLeftCommand.cs
@@ -12,9 +12,9 @@ public class TurnLeftCommand : ICommand
 {
     private readonly ITroop troop;
 
-    private readonly float degrees;
+    private readonly int degrees;
 
-    public TurnLeftCommand(Infantry _troop, float _degrees)
+    public TurnLeftCommand(Infantry _troop, int _degrees)
     {
         troop = _troop;
         degrees = _degrees;
